- Embedding and POC with OpenAI is done
